reset challenge to start
