+++ sessions/backend-update-and-delete/products/components/ProductForm/index.js
@@ -1,17 +1,36 @@
 import { StyledForm, StyledHeading, StyledLabel } from "./ProductForm.styled";
 import { StyledButton } from "../Button/Button.styled";
+import useSWR from "swr";
 
-export default function ProductForm({ onSubmit, heading }) {
+export default function ProductForm() {
+  const products = useSWR("/api/products");
+
   async function handleSubmit(event) {
     event.preventDefault();
-    onSubmit(event);
 
-    event.target.reset();
+    const formData = new FormData(event.target);
+    const productData = Object.fromEntries(formData);
+
+    const response = await fetch("/api/products", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(productData),
+    });
+
+    if (response.ok) {
+      await response.json();
+      products.mutate();
+      event.target.reset();
+    } else {
+      console.error(response.status);
+    }
   }
 
   return (
     <StyledForm onSubmit={handleSubmit}>
-      <StyledHeading>{heading}</StyledHeading>
+      <StyledHeading>Add a new Fish</StyledHeading>
       <StyledLabel htmlFor="name">
         Name:
         <input type="text" id="name" name="name" />
