+++ sessions/backend-update-and-delete/products/pages/api/products/[id].js
@@ -14,18 +14,4 @@ export default async function handler(request, response) {
 
     response.status(200).json(product);
   }
-
-  if (request.method === "PUT") {
-    const productToUpdate = await Product.findByIdAndUpdate(id, {
-      $set: request.body,
-    });
-    return response.status(200).json(productToUpdate);
-  }
-
-  if (request.method === "DELETE") {
-    await Product.findByIdAndDelete(id);
-    return response
-      .status(200)
-      .json({ status: "Product successfully deleted." });
-  }
 }
