+++ sessions/backend-update-and-delete/products/pages/index.js
@@ -1,7 +1,6 @@
 import ProductList from "../components/ProductList";
 import styled from "styled-components";
 import ProductForm from "../components/ProductForm";
-import useSWR from "swr";
 
 const Heading = styled.h1`
   text-align: center;
@@ -9,28 +8,6 @@ const Heading = styled.h1`
 `;
 
 export default function HomePage() {
-  const products = useSWR("/api/products");
-
-  async function handleAddProduct(event) {
-    const formData = new FormData(event.target);
-    const productData = Object.fromEntries(formData);
-
-    const response = await fetch("/api/products", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(productData),
-    });
-
-    if (response.ok) {
-      await response.json();
-      products.mutate();
-    } else {
-      console.error(response.status);
-    }
-  }
-
   return (
     <>
       <Heading>
@@ -39,7 +16,7 @@ export default function HomePage() {
         </span>
         Fish Shop
       </Heading>
-      <ProductForm onSubmit={handleAddProduct} heading="Add a new Fish" />
+      <ProductForm />
       <hr />
       <ProductList />
     </>
